Create .gitignore
Created a .gitignore file that will ensure git will not check in the .vs configuration folder from visual studio, nor the default out folder where the project is built to.

It may be necessary to add additional things to the gitignore file for use with other IDEs or other operating systems. This file will be updated as needed in the classroom.